ref(utils): Move progress bar code to own submodule
